[Silver III] Title: 2×n 타일링, Time: 124 ms, Memory: 14320 KB -BaekjoonHub

+++ 백준/Silver/11726. 2×n 타일링/2×n 타일링.java
@@ -8,7 +8,7 @@ public class Main {
         StringTokenizer st = new StringTokenizer(br.readLine());
         int n = Integer.parseInt(st.nextToken());
 
-        long[] dp = new long[1001];
+        int[] dp = new int[1001];
         for(int i=0; i<=n; i++){
             if(i==1) dp[i] = 1;
             if(i==2) dp[i] = 2;
